requirements.txt
invoke==2.2.0
matplotlib==3.10.0
Pillow==11.1.0
timm==1.0.13
torch==2.5.1
torchvision==0.20.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
loguru==0.7.3
hydra-core==1.3.2
python-dotenv==1.0.1
pandas==2.2.3
wandb==0.19.4
pytorch-lightning==2.5.0.post0
uvicorn==0.34.0
fastapi==0.115.6
python-multipart==0.0.20
pytest
coverage
